+++ examples/circosPlot-examples.R
@@ -34,9 +34,31 @@ circosPlot(nutrimouse.sgccda, cutoff = 0.7, line = TRUE, ncol.legend = 1,
 ## custom line legends
 circosPlot(nutrimouse.sgccda, cutoff = 0.7, line = TRUE, ncol.legend = 2,
            size.legend = 1.1, showIntraLinks = TRUE)
-
+par(mfrow=c(1,1))
 
 ## adjust feature and block names radially
 circosPlot(nutrimouse.sgccda, cutoff = 0.7, size.legend = 1.1)
 circosPlot(nutrimouse.sgccda, cutoff = 0.7, size.legend = 1.1, var.adj = 0.8, block.labels.adj = -0.5)
+## ---  example using breast.TCGA data
+data("breast.TCGA")
+data = list(mrna = breast.TCGA$data.train$mrna, 
+            mirna = breast.TCGA$data.train$mirna, 
+            protein = breast.TCGA$data.train$protein)
+list.keepX = list(mrna = rep(20, 2), mirna = rep(10,2), protein = c(10, 2))
+
+TCGA.block.splsda = block.splsda(X = data, 
+                             Y =breast.TCGA$data.train$subtype,
+                             ncomp = 2, keepX = list.keepX, 
+                             design = 'full')
+circosPlot(TCGA.block.splsda, cutoff = 0.7, line=TRUE)
+data("breast.TCGA")
+data = list(mrna = breast.TCGA$data.train$mrna, mirna = breast.TCGA$data.train$mirna)
+list.keepX = list(mrna = rep(20, 2), mirna = rep(10,2))
+list.keepY = c(rep(10, 2))
+
+TCGA.block.spls = block.spls(X = data, 
+                             Y = breast.TCGA$data.train$protein,
+                             ncomp = 2, keepX = list.keepX, 
+                             keepY = list.keepY, design = 'full')
+circosPlot(TCGA.block.spls, group = breast.TCGA$data.train$subtype, cutoff = 0.7, Y.name = 'protein')
 }