+++ NEWS.md
@@ -4,6 +4,7 @@
 
 * `plotIndiv.pca` default shape set to `16`
 * `circosPlot` & `network` now support blocks with similar feature names 
+* `circosPlot` now has methods for block.spls objects
 
 ### bug fixes